Replace hard coded Status Codes with HttpStatusCode (#219)
* replaced int with HTTPStatusCode

* replaced int with httpstatus code in tests

* changed int to httpstatuscode in request validator unit tests

+++ DataGateway.Service/Resolvers/Sql Query Structures/SqlUpdateQueryStructure.cs
@@ -54,7 +54,7 @@ namespace Azure.DataGateway.Service.Resolvers
             {
                 throw new DatagatewayException(
                     message: "Update mutation does not update any values",
-                    statusCode: (int)HttpStatusCode.BadRequest,
+                    statusCode: HttpStatusCode.BadRequest,
                     subStatusCode: DatagatewayException.SubStatusCodes.BadRequest);
             }
         }
